refactor: split oversized Vue components
