+++ frontend/src/components/album/MediaItem.vue
@@ -398,206 +398,4 @@ function onVideoKey(e: KeyboardEvent) {
   </div>
 </template>
 
-<style lang="scss" scoped>
-.media-item {
-  container-type: size;
-  background: transparent;
-  width: 100%;
-  height: 100%;
-
-  &.selectable {
-    cursor: grab;
-  }
-
-  &.selectable:active {
-    cursor: grabbing;
-  }
-
-  &.focused {
-    outline: 4px solid var(--q-primary);
-    outline-offset: -4px;
-    box-shadow: 0 0 0 2px rgba(0, 99, 209, 0.35);
-  }
-
-  &:focus-visible {
-    outline: 2px solid var(--q-primary);
-    outline-offset: -2px;
-  }
-}
-
-.fit-cover {
-  object-fit: cover;
-}
-
-.fit-contain {
-  object-fit: contain;
-}
-
-.video-playing {
-  object-fit: contain;
-  background: black; /* letterbox - intentionally pure black, not theme-adaptive */
-}
-
-// Button reset - <button> replaces <div> for a11y but needs native chrome stripped.
-// Don't use all:unset (it kills Quasar utility classes on the same element).
-.play-overlay {
-  background: none;
-  border: none;
-  padding: 0;
-  font: inherit;
-  color: inherit;
-}
-
-// Play button - scales with container so it looks right in both
-// full-size page cells and the small unused-photos tray.
-
-.play-icon {
-  // Scale with the smaller container dimension, capped at 3.5rem
-  --size: min(3.5rem, 40cqmin);
-  width: var(--size);
-  height: var(--size);
-  border-radius: 50%;
-  background: rgba(0, 0, 0, 0.5);
-  transition: background var(--duration-fast);
-
-  :deep(.q-icon) {
-    font-size: calc(var(--size) * 0.55);
-    color: white;
-  }
-
-  .play-overlay:hover & {
-    background: rgba(0, 0, 0, 0.7);
-  }
-}
-
-// Frame selection toolbar - appears above native video controls
-.frame-bar {
-  position: absolute;
-  bottom: 3.5rem; // clear native controls
-  left: 0;
-  right: 0;
-  width: fit-content;
-  margin: 0 auto;
-  gap: var(--gap-sm);
-  background: rgba(0, 0, 0, 0.75);
-  backdrop-filter: blur(10px);
-  border-radius: var(--radius-full);
-  padding: var(--gap-sm);
-}
-
-.frame-step-btn {
-  width: 2rem;
-  height: 2rem;
-  border: none;
-  border-radius: 50%;
-  background: transparent;
-  color: rgba(255, 255, 255, 0.8);
-  cursor: pointer;
-  transition: background var(--duration-fast);
-
-  :deep(.q-icon) {
-    font-size: var(--type-lg);
-  }
-
-  &:hover {
-    background: rgba(255, 255, 255, 0.15);
-  }
-
-  &:focus-visible {
-    outline: 2px solid white;
-    outline-offset: 2px;
-  }
-}
-
-.set-frame-btn {
-  gap: var(--gap-sm-md);
-  border: none;
-  border-radius: var(--radius-full);
-  /* always on dark video backdrop - not theme-adaptive */
-  background: white;
-  color: #111;
-  font-size: var(--type-sm);
-  font-weight: 600;
-  padding: var(--gap-sm-md) var(--gap-lg) var(--gap-sm-md) var(--gap-md-lg);
-  cursor: pointer;
-  transition:
-    background var(--duration-fast),
-    transform var(--duration-fast);
-
-  /* always on dark backdrop - light-mode-only button */
-  &:hover {
-    background: color-mix(in srgb, white 85%, black);
-  }
-
-  &:active {
-    transform: scale(0.96);
-  }
-
-  &:focus-visible {
-    outline: 2px solid var(--q-primary);
-    outline-offset: 2px;
-  }
-}
-
-.quality-badge {
-  --size: min(2.25rem, 34cqmin);
-  position: absolute;
-  bottom: var(--gap-md);
-  right: var(--gap-md);
-  width: var(--size);
-  height: var(--size);
-  border: none;
-  border-radius: 50%;
-  pointer-events: auto;
-  z-index: 2;
-  cursor: pointer;
-  box-shadow: 0 0.125rem 0.5rem color-mix(in srgb, black 35%, transparent);
-
-  &.caution {
-    background: color-mix(in srgb, var(--q-warning) 80%, black);
-  }
-
-  &.warning {
-    background: color-mix(in srgb, var(--danger) 80%, black);
-  }
-
-  :deep(.q-icon) {
-    font-size: calc(var(--size) * 0.55);
-    color: white;
-  }
-
-  &:focus-visible {
-    outline: 0.125rem solid white;
-    outline-offset: 0.125rem;
-  }
-}
-
-
-@media print {
-  .play-overlay,
-  .frame-bar,
-  .quality-badge {
-    display: none !important;
-  }
-
-}
-
-@media (pointer: coarse) {
-  .frame-step-btn {
-    width: 2.75rem;
-    height: 2.75rem;
-  }
-
-  .set-frame-btn {
-    padding: var(--gap-md) var(--gap-lg) var(--gap-md) var(--gap-md-lg);
-  }
-}
-
-@media (prefers-reduced-motion: reduce) {
-  .play-icon,
-  .frame-step-btn,
-  .set-frame-btn {
-    transition: none;
-  }
-}
-</style>
+<style lang="scss" scoped src="./MediaItem.scss"></style>
