+++ frontend/src/components/editor/MediaPanel.scss
@@ -1,0 +1,383 @@
+.media-panel {
+  display: flex;
+  flex-direction: column;
+  gap: var(--gap-sm);
+  padding: var(--gap-md) var(--gap-lg);
+}
+
+.hidden-input {
+  position: absolute;
+  width: 0.0625rem;
+  height: 0.0625rem;
+  opacity: 0;
+  pointer-events: none;
+}
+
+.sr-only {
+  position: absolute;
+  width: 0.0625rem;
+  height: 0.0625rem;
+  margin: -0.0625rem;
+  padding: 0;
+  overflow: hidden;
+  clip: rect(0, 0, 0, 0);
+  white-space: nowrap;
+  border: 0;
+}
+
+.quality-section {
+  display: flex;
+  flex-direction: column;
+  gap: var(--gap-sm);
+}
+
+.quality-controls {
+  display: flex;
+  flex-direction: column;
+  gap: var(--gap-md);
+}
+
+.quality-header {
+  display: flex;
+  flex-wrap: wrap;
+  align-items: center;
+  justify-content: space-between;
+  column-gap: var(--gap-sm);
+  row-gap: var(--gap-xs);
+}
+
+.quality-section-title {
+  margin: 0;
+  font-size: var(--type-xs);
+  font-weight: 500;
+  color: var(--text-muted);
+}
+
+.quality-chip {
+  all: unset;
+  display: inline-flex;
+  align-items: center;
+  gap: var(--gap-xs);
+  min-height: 2.5rem;
+  padding: var(--gap-sm) var(--gap-md);
+  border-radius: var(--radius-xs);
+  font-family: var(--font-ui);
+  font-size: var(--type-sm);
+  font-weight: 500;
+  color: var(--text);
+  cursor: pointer;
+  transition:
+    background var(--duration-fast),
+    color var(--duration-fast);
+
+  &.warning .chip-icon {
+    color: var(--q-negative);
+  }
+
+  &.caution .chip-icon {
+    color: var(--q-warning);
+  }
+
+  &.all-clear {
+    color: var(--text-muted);
+    font-weight: 400;
+    cursor: default;
+  }
+
+  &.all-clear .chip-icon {
+    color: inherit;
+  }
+
+  &:not(.all-clear):hover {
+    background: color-mix(in srgb, var(--text) 6%, transparent);
+  }
+
+  &:focus-visible {
+    outline: 0.125rem solid var(--q-primary);
+    outline-offset: 0.125rem;
+  }
+}
+
+.media-cta {
+  all: unset;
+  box-sizing: border-box;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  gap: var(--gap-sm);
+  width: 100%;
+  min-height: 2.75rem;
+  padding: 0 var(--gap-md-lg);
+  border: 1px solid var(--q-primary);
+  border-radius: var(--radius-sm);
+  background: transparent;
+  color: var(--q-primary);
+  font-family: var(--font-ui);
+  font-size: var(--type-sm);
+  font-weight: 500;
+  cursor: pointer;
+  transition:
+    background var(--duration-fast),
+    border-color var(--duration-fast),
+    color var(--duration-fast);
+
+  &:hover:not(:disabled) {
+    background: var(--q-primary);
+    color: var(--bg);
+  }
+
+  &:focus-visible {
+    outline: 0.125rem solid var(--q-primary);
+    outline-offset: 0.125rem;
+  }
+
+  &:disabled {
+    cursor: default;
+    opacity: 0.5;
+  }
+}
+
+.media-cta.primary {
+  background: var(--q-primary);
+  color: var(--bg);
+
+  &:hover:not(:disabled) {
+    background: color-mix(in srgb, var(--q-primary) 88%, var(--text));
+    border-color: color-mix(in srgb, var(--q-primary) 88%, var(--text));
+    color: var(--bg);
+  }
+}
+
+.media-cta.subtle {
+  border-color: transparent;
+  color: var(--text-muted);
+
+  &:hover:not(:disabled) {
+    background: transparent;
+    color: var(--text);
+    border-color: var(--text-faint);
+  }
+}
+
+.import-cta {
+  display: flex;
+  align-items: stretch;
+  width: 100%;
+}
+
+.import-cta-main {
+  flex: 1;
+  width: auto;
+
+  &.has-trailing {
+    border-start-end-radius: 0;
+    border-end-end-radius: 0;
+  }
+}
+
+.import-cta-trigger {
+  all: unset;
+  box-sizing: border-box;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  min-width: 2.75rem;
+  padding: 0 var(--gap-sm);
+  border: 1px solid var(--q-primary);
+  border-inline-start: 1px solid color-mix(in srgb, var(--bg) 35%, transparent);
+  border-start-start-radius: 0;
+  border-end-start-radius: 0;
+  border-start-end-radius: var(--radius-sm);
+  border-end-end-radius: var(--radius-sm);
+  background: var(--q-primary);
+  color: var(--bg);
+  cursor: pointer;
+  transition:
+    background var(--duration-fast),
+    border-color var(--duration-fast);
+
+  &:hover:not(:disabled) {
+    background: color-mix(in srgb, var(--q-primary) 88%, var(--text));
+    border-color: color-mix(in srgb, var(--q-primary) 88%, var(--text));
+  }
+
+  &:focus-visible {
+    outline: 0.125rem solid var(--q-primary);
+    outline-offset: 0.125rem;
+  }
+
+  &:disabled {
+    cursor: default;
+    opacity: 0.5;
+  }
+}
+
+.media-helper {
+  margin: 0;
+  font-size: var(--type-xs);
+  line-height: 1.5;
+  color: var(--text-muted);
+}
+
+.media-error {
+  margin: 0;
+  font-size: var(--type-xs);
+  line-height: 1.4;
+  color: var(--q-negative);
+}
+
+.selected-section {
+  display: flex;
+  flex-direction: column;
+  gap: var(--gap-sm);
+}
+
+.replace-swap {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: var(--gap-sm);
+}
+
+.swap-cell {
+  flex: 1;
+  min-width: 3rem;
+  max-width: 6rem;
+  aspect-ratio: 1;
+  border-radius: var(--radius-sm);
+  overflow: hidden;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  background: color-mix(in srgb, var(--text) 4%, transparent);
+  color: var(--text-muted);
+}
+
+.swap-cell.current img {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  display: block;
+}
+
+.swap-cell.target {
+  all: unset;
+  box-sizing: border-box;
+  flex: 1;
+  min-width: 3rem;
+  max-width: 6rem;
+  aspect-ratio: 1;
+  border-radius: var(--radius-sm);
+  border: 1px solid var(--text-faint);
+  background: transparent;
+  color: var(--text-muted);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
+  transition:
+    border-color var(--duration-fast),
+    background var(--duration-fast),
+    color var(--duration-fast);
+
+  &:hover:not(:disabled) {
+    border-color: var(--q-primary);
+    color: var(--q-primary);
+    background: color-mix(in srgb, var(--q-primary) 8%, transparent);
+  }
+
+  &:focus-visible {
+    outline: 0.125rem solid var(--q-primary);
+    outline-offset: 0.125rem;
+  }
+
+  &:disabled {
+    cursor: default;
+    opacity: 0.5;
+  }
+}
+
+.swap-arrow {
+  color: var(--text-faint);
+  flex-shrink: 0;
+}
+
+.cta-menu {
+  display: grid;
+  gap: var(--gap-xs);
+  min-width: 13rem;
+  padding: var(--gap-xs);
+}
+
+.cta-menu-item {
+  all: unset;
+  box-sizing: border-box;
+  display: flex;
+  align-items: center;
+  gap: var(--gap-sm);
+  width: 100%;
+  min-height: 2.75rem;
+  padding: 0 var(--gap-md);
+  border-radius: var(--radius-xs);
+  color: var(--text);
+  cursor: pointer;
+  font-family: var(--font-ui);
+  font-size: var(--type-sm);
+
+  &:hover:not(:disabled),
+  &:focus-visible {
+    background: color-mix(in srgb, var(--q-primary) 10%, transparent);
+  }
+
+  &:disabled {
+    cursor: default;
+    opacity: 0.45;
+  }
+}
+
+.import-dialog {
+  width: min(24rem, 90vw);
+  border-radius: var(--radius-sm);
+  padding: var(--gap-lg);
+  display: grid;
+  gap: var(--gap-md);
+}
+
+.import-dialog-header {
+  display: flex;
+  align-items: center;
+  gap: var(--gap-sm);
+  color: var(--text);
+}
+
+.import-dialog-title {
+  margin: 0;
+  font-family: var(--font-ui);
+  font-size: var(--type-md);
+  font-weight: 600;
+  line-height: 1.3;
+}
+
+.import-dialog-progress {
+  margin: 0;
+}
+
+.import-dialog-actions {
+  display: flex;
+  justify-content: flex-end;
+  gap: var(--gap-sm);
+}
+
+.import-dialog-actions .media-cta {
+  width: auto;
+  padding: 0 var(--gap-md-lg);
+}
+
+@media (prefers-reduced-motion: reduce) {
+  .quality-chip,
+  .media-cta,
+  .import-cta-trigger,
+  .swap-cell.target {
+    transition: none;
+  }
+}
